[#47017] Replace Versions::TableCell by ViewComponent
https://community.openproject.org/work_packages/47017
