+++ modules/backlogs/lib/open_project/backlogs/patches/versions/row_component_patch.rb
@@ -1,0 +1,44 @@
+#-- copyright
+# OpenProject is an open source project management software.
+# Copyright (C) 2012-2023 the OpenProject GmbH
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License version 3.
+#
+# OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+# Copyright (C) 2006-2013 Jean-Philippe Lang
+# Copyright (C) 2010-2013 the ChiliProject Team
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+# See COPYRIGHT and LICENSE files for more details.
+#++
+
+module OpenProject::Backlogs::Patches::Versions::RowComponentPatch
+  def button_links
+    (super + [backlogs_edit_link]).compact
+  end
+
+  private
+
+  def backlogs_edit_link
+    return if version.project == table.project || !table.project.module_enabled?("backlogs")
+
+    link_to_if_authorized '',
+                          { controller: '/versions', action: 'edit', id: version, project_id: table.project.id },
+                          class: 'icon icon-edit',
+                          title: t(:button_edit)
+  end
+end
